requirements.txt
torch
transformers
pillow

